feat(arch): 既存クラスを 4 層パッケージへ移動 (N4 / A-2) (#272)
task / security / tenant / user feature の既存クラスを
A-1 (#86) で配置した 4 層パッケージ雛形へ移動する。

移動マッピング:
- task/Task.java              → task/adapter/persistence/Task.java
- task/Priority.java          → task/domain/Priority.java
- task/TaskStatus.java        → task/domain/TaskStatus.java
- task/Visibility.java        → task/domain/Visibility.java
- security/SecurityConfig.java         → security/adapter/web/SecurityConfig.java
- security/TasksAuthenticationToken.java → security/adapter/web/TasksAuthenticationToken.java
- security/TasksJwtAuthenticationConverter.java → security/adapter/web/TasksJwtAuthenticationConverter.java
- security/TasksPrincipal.java → security/domain/TasksPrincipal.java
- tenant/Tenant.java          → tenant/adapter/persistence/Tenant.java
- tenant/TenantPlan.java      → tenant/domain/TenantPlan.java
- tenant/TenantStatus.java    → tenant/domain/TenantStatus.java
- user/User.java              → user/adapter/persistence/User.java
- user/UserRepository.java    → user/adapter/persistence/UserRepository.java

移動に伴う変更:
- package 宣言・import を新パスに合わせて修正
- task/domain/TaskAuthorizationDomainService の import 更新
- 全テストファイルの import 更新(package は変更なし)
- user/adapter/persistence/package-info.java に @NamedInterface 追加
  (security feature が User / UserRepository を参照するため)
- 旧パスのファイルを空 compilation unit に差し替え(git rename 検知)
