+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/security/adapter/web/TasksAuthenticationToken.java
@@ -1,0 +1,32 @@
+package xyz.dgz48.tasks.webapi.security.adapter.web;
+
+import java.util.Collection;
+import org.jspecify.annotations.Nullable;
+import org.springframework.security.authentication.AbstractAuthenticationToken;
+import org.springframework.security.core.GrantedAuthority;
+import xyz.dgz48.tasks.webapi.security.domain.TasksPrincipal;
+
+public class TasksAuthenticationToken extends AbstractAuthenticationToken {
+
+  private static final long serialVersionUID = 1L;
+
+  private final TasksPrincipal principal;
+
+  public TasksAuthenticationToken(
+      TasksPrincipal principal, Collection<? extends GrantedAuthority> authorities) {
+    super(authorities);
+    this.principal = principal;
+    setAuthenticated(true);
+  }
+
+  @Override
+  @Nullable
+  public Object getCredentials() {
+    return null;
+  }
+
+  @Override
+  public TasksPrincipal getPrincipal() {
+    return principal;
+  }
+}
