+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/task/Priority.java
@@ -1,7 +1,1 @@
 package xyz.dgz48.tasks.webapi.task;
-
-public enum Priority {
-  HIGH,
-  MEDIUM,
-  LOW
-}
