+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/task/TaskStatus.java
@@ -1,8 +1,1 @@
 package xyz.dgz48.tasks.webapi.task;
-
-public enum TaskStatus {
-  NOT_STARTED,
-  IN_PROGRESS,
-  DONE,
-  ON_HOLD
-}
