+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/task/domain/TaskStatus.java
@@ -1,0 +1,8 @@
+package xyz.dgz48.tasks.webapi.task.domain;
+
+public enum TaskStatus {
+  NOT_STARTED,
+  IN_PROGRESS,
+  DONE,
+  ON_HOLD
+}
