+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/task/domain/Visibility.java
@@ -1,0 +1,7 @@
+package xyz.dgz48.tasks.webapi.task.domain;
+
+public enum Visibility {
+  TENANT,
+  STAKEHOLDERS,
+  PRIVATE
+}
