+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/tenant/TenantPlan.java
@@ -1,7 +1,1 @@
 package xyz.dgz48.tasks.webapi.tenant;
-
-public enum TenantPlan {
-  FREE,
-  STANDARD,
-  ENTERPRISE
-}
