+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/tenant/TenantStatus.java
@@ -1,7 +1,1 @@
 package xyz.dgz48.tasks.webapi.tenant;
-
-public enum TenantStatus {
-  ACTIVE,
-  SUSPENDED,
-  DELETED
-}
