+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/tenant/domain/TenantPlan.java
@@ -1,0 +1,7 @@
+package xyz.dgz48.tasks.webapi.tenant.domain;
+
+public enum TenantPlan {
+  FREE,
+  STANDARD,
+  ENTERPRISE
+}
