+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/tenant/domain/TenantStatus.java
@@ -1,0 +1,7 @@
+package xyz.dgz48.tasks.webapi.tenant.domain;
+
+public enum TenantStatus {
+  ACTIVE,
+  SUSPENDED,
+  DELETED
+}
