+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/user/UserRepository.java
@@ -1,8 +1,1 @@
 package xyz.dgz48.tasks.webapi.user;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-  Optional<User> findByOidcSub(String oidcSub);
-}
