+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/user/adapter/persistence/UserRepository.java
@@ -1,0 +1,8 @@
+package xyz.dgz48.tasks.webapi.user.adapter.persistence;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+  Optional<User> findByOidcSub(String oidcSub);
+}
