+++ webapi/src/main/java/xyz/dgz48/tasks/webapi/user/adapter/persistence/package-info.java
@@ -1,4 +1,6 @@
+@NamedInterface
 @NullMarked
 package xyz.dgz48.tasks.webapi.user.adapter.persistence;
 
 import org.jspecify.annotations.NullMarked;
+import org.springframework.modulith.NamedInterface;
