+++ webapi/src/test/java/xyz/dgz48/tasks/webapi/security/TasksAuthenticationTokenTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import xyz.dgz48.tasks.webapi.security.adapter.web.TasksAuthenticationToken;
+import xyz.dgz48.tasks.webapi.security.domain.TasksPrincipal;
 
 class TasksAuthenticationTokenTest {
